+++ FlightBookingApp/src/utils/firebase/FirebaseFunctions.js
@@ -1,6 +1,11 @@
 import { ToastAndroid } from 'react-native'
 import firestore from '@react-native-firebase/firestore';
 import auth from '@react-native-firebase/auth'
+import { GoogleSignin } from '@react-native-google-signin/google-signin'
+
+GoogleSignin.configure({
+  webClientId: '43375129789-19d3mo4bim7cgmt6d7co7lr44doerqti.apps.googleusercontent.com'
+});
 
 export const loginAuth = async (navigation, email, pwd) =>{
     try{
@@ -29,20 +34,20 @@ export const loginAuth = async (navigation, email, pwd) =>{
     }
 }
 
-// export const getUserdata = async (props) =>{
-//     let infoUser = ""
-//     try{
-//       await firestore()
-//         .collection('Users')
-//         .doc(props.user.uid)
-//         .get()
-//         .then(res2=>{
-//             alert("Welcome "+res2._data.name)
-//             infoUser = props.user.uid
-//             console.log("Soy el uid" + infoUser)
-//           })
-//     } catch (e){
-//       console.log('Error: ', e)
-//     }
-//   return infoUser
-// }
+export const SignInWithGoogle = async (navigation) => {
+  // console.log('Hola')
+  // GoogleSignin.signOut()
+  let infoUser = ""
+  // Get the users ID token
+  
+  const { idToken } = await GoogleSignin.signIn();
+  // Create a Google credential with the token
+  
+  const googleCredential = auth.GoogleAuthProvider.credential(idToken);
+  // Sign-in the user with the credential
+  
+  await auth().signInWithCredential(googleCredential).then((resp)=>{
+    infoUser = resp.user.uid
+  });
+  navigation.navigate('My Flights', infoUser)
+}
